[fix](cast) fix implicit casts in vertical_segment_writer

+++ be/src/olap/rowset/segment_v2/vertical_segment_writer.cpp
@@ -76,6 +76,8 @@
 
 namespace doris::segment_v2 {
 
+#include "common/compile_check_begin.h"
+
 using namespace ErrorCode;
 
 static const char* k_segment_magic = "D0R1";
@@ -114,7 +116,7 @@ VerticalSegmentWriter::VerticalSegmentWriter(io::FileWriter* file_writer, uint32
     for (size_t cid = 0; cid < _num_sort_key_columns; ++cid) {
         const auto& column = _tablet_schema->column(cid);
         _key_coders.push_back(get_key_coder(column.type()));
-        _key_index_size.push_back(column.index_length());
+        _key_index_size.push_back(static_cast<uint16_t>(column.index_length()));
     }
     // encode the sequence id into the primary key index
     if (_is_mow()) {
@@ -135,7 +137,7 @@ VerticalSegmentWriter::VerticalSegmentWriter(io::FileWriter* file_writer, uint32
             for (auto cid : _tablet_schema->cluster_key_uids()) {
                 const auto& column = _tablet_schema->column_by_uid(cid);
                 _key_coders.push_back(get_key_coder(column.type()));
-                _key_index_size.push_back(column.index_length());
+                _key_index_size.push_back(static_cast<uint16_t>(column.index_length()));
             }
         }
     }
@@ -149,7 +151,7 @@ void VerticalSegmentWriter::_init_column_meta(ColumnMetaPB* meta, uint32_t colum
                                               const TabletColumn& column) {
     meta->set_column_id(column_id);
     meta->set_type(int(column.type()));
-    meta->set_length(column.length());
+    meta->set_length(static_cast<int32_t>(column.length()));
     meta->set_encoding(DEFAULT_ENCODING);
     meta->set_compression(_opts.compression_type);
     meta->set_is_nullable(column.is_nullable());
@@ -203,8 +205,8 @@ Status VerticalSegmentWriter::_create_column_writer(uint32_t cid, const TabletCo
             return Status::NotSupported("Do not support ngram bloom filter for bf_size: ",
                                         gram_bf_size);
         }
-        opts.gram_size = gram_size;
-        opts.gram_bf_size = gram_bf_size;
+        opts.gram_size = static_cast<uint8_t>(gram_size);
+        opts.gram_bf_size = static_cast<uint16_t>(gram_bf_size);
     }
 
     opts.need_bitmap_index = column.has_bitmap_index();
@@ -352,8 +354,8 @@ void VerticalSegmentWriter::_serialize_block_to_row_column(vectorized::Block& bl
             std::unordered_set<int> row_store_cids_set(_tablet_schema->row_columns_uids().begin(),
                                                        _tablet_schema->row_columns_uids().end());
             vectorized::JsonbSerializeUtil::block_to_jsonb(
-                    *_tablet_schema, block, *row_store_column, _tablet_schema->num_columns(),
-                    serdes, row_store_cids_set);
+                    *_tablet_schema, block, *row_store_column,
+                    static_cast<int>(_tablet_schema->num_columns()), serdes, row_store_cids_set);
             break;
         }
     }
@@ -373,9 +375,9 @@ Status VerticalSegmentWriter::_probe_key_for_mow(
     RowLocation loc;
     // save rowset shared ptr so this rowset wouldn't delete
     RowsetSharedPtr rowset;
-    auto st = _tablet->lookup_row_key(key, _tablet_schema.get(), have_input_seq_column,
-                                      specified_rowsets, &loc, _mow_context->max_version,
-                                      segment_caches, &rowset);
+    auto st = _tablet->lookup_row_key(
+            key, _tablet_schema.get(), have_input_seq_column, specified_rowsets, &loc,
+            static_cast<uint32_t>(_mow_context->max_version), segment_caches, &rowset);
     if (st.is<KEY_NOT_FOUND>()) {
         if (!have_delete_sign) {
             RETURN_IF_ERROR(not_found_cb());
@@ -418,7 +420,7 @@ Status VerticalSegmentWriter::_probe_key_for_mow(
         // for this row, we need to ensure that each column is aligned
         _mow_context->delete_bitmap->add(
                 {_opts.rowset_ctx->rowset_id, _segment_id, DeleteBitmap::TEMP_VERSION_COMMON},
-                segment_pos);
+                static_cast<uint32_t>(segment_pos));
         ++stats.num_rows_deleted;
     } else {
         _mow_context->delete_bitmap->add(
@@ -562,7 +564,8 @@ Status VerticalSegmentWriter::_append_block_with_partial_content(RowsInBlock& da
         auto not_found_cb = [&]() {
             return _opts.rowset_ctx->partial_update_info->handle_new_key(
                     *_tablet_schema, [&]() -> std::string {
-                        return data.block->dump_one_line(block_pos, _num_sort_key_columns);
+                        return data.block->dump_one_line(block_pos,
+                                                         static_cast<int>(_num_sort_key_columns));
                     });
         };
         auto update_read_plan = [&](const RowLocation& loc) {
@@ -733,7 +736,8 @@ Status VerticalSegmentWriter::_append_block_with_flexible_partial_content(
     // 8. encode and write all non-primary key columns(including sequence column if exists)
     for (auto cid = _tablet_schema->num_key_columns(); cid < _tablet_schema->num_columns(); cid++) {
         RETURN_IF_ERROR(_olap_data_convertor->set_source_content_with_specifid_column(
-                full_block.get_by_position(cid), data.row_pos, data.num_rows, cid));
+                full_block.get_by_position(cid), data.row_pos, data.num_rows,
+                static_cast<uint32_t>(cid)));
         auto [status, column] = _olap_data_convertor->convert_column_data(cid);
         if (!status.ok()) {
             return status;
@@ -839,7 +843,8 @@ Status VerticalSegmentWriter::_generate_flexible_read_plan(
             return _opts.rowset_ctx->partial_update_info->handle_new_key(
                     *_tablet_schema,
                     [&]() -> std::string {
-                        return data.block->dump_one_line(block_pos, _num_sort_key_columns);
+                        return data.block->dump_one_line(block_pos,
+                                                         static_cast<int>(_num_sort_key_columns));
                     },
                     &skip_bitmap);
         };
@@ -938,7 +943,7 @@ Status VerticalSegmentWriter::_append_block_with_variant_subcolumns(RowsInBlock&
                 continue;
             }
             CHECK(entry->data.is_finalized());
-            int current_column_id = column_id++;
+            uint32_t current_column_id = static_cast<uint32_t>(column_id++);
             TabletColumn tablet_column = generate_column_info(entry);
             vectorized::schema_util::inherit_column_attributes(*parent_column, tablet_column,
                                                                _flush_schema);
@@ -1433,7 +1438,7 @@ Status VerticalSegmentWriter::_write_footer() {
 
     faststring fixed_buf;
     // footer's size
-    put_fixed32_le(&fixed_buf, footer_buf.size());
+    put_fixed32_le(&fixed_buf, static_cast<uint32_t>(footer_buf.size()));
     // footer's checksum
     uint32_t checksum = crc32c::Value(footer_buf.data(), footer_buf.size());
     put_fixed32_le(&fixed_buf, checksum);
@@ -1489,4 +1494,7 @@ inline bool VerticalSegmentWriter::_is_mow() {
 inline bool VerticalSegmentWriter::_is_mow_with_cluster_key() {
     return _is_mow() && !_tablet_schema->cluster_key_uids().empty();
 }
+
+#include "common/compile_check_end.h"
+
 } // namespace doris::segment_v2
